+++ lib/td_auth/accounts/user.ex
@@ -29,7 +29,7 @@ defmodule TdAuth.Accounts.User do
 
   def changeset(params), do: changeset(%__MODULE__{}, params)
 
-  def changeset(%__MODULE__{} = user, params) do
+  def changeset(%__MODULE__{} = user, params, keep_groups \\ false) do
     user
     |> cast(params, [:user_name, :role, :is_admin, :email, :full_name])
     |> cast(params, [:password, :old_password], empty_values: [])
@@ -38,7 +38,7 @@ defmodule TdAuth.Accounts.User do
     |> validate_old_password()
     |> put_pass_hash()
     |> update_change(:user_name, &String.downcase/1)
-    |> put_groups(params)
+    |> put_groups(user.groups, keep_groups, params)
     |> put_role()
     |> unique_constraint(:user_name)
   end
@@ -66,11 +66,15 @@ defmodule TdAuth.Accounts.User do
 
   defp put_role(%Changeset{} = changeset), do: changeset
 
-  defp put_groups(%Changeset{valid?: true} = changeset, %{"groups" => groups}) do
+  defp put_groups(%Changeset{valid?: true} = changeset, _existing_groups, false, %{"groups" => groups}) do
     put_assoc(changeset, :groups, groups)
   end
 
-  defp put_groups(%Changeset{} = changeset, _), do: changeset
+  defp put_groups(%Changeset{valid?: true} = changeset, existing_groups, true, %{"groups" => groups}) do
+    put_assoc(changeset, :groups, existing_groups ++ groups)
+  end
+
+  defp put_groups(%Changeset{} = changeset, _existing_groups, _, _), do: changeset
 
   defp put_pass_hash(%Ecto.Changeset{valid?: true, changes: %{password: password}} = changeset) do
     hash = Bcrypt.add_hash(password)