+++ lib/td_auth_web/controllers/session_controller.ex
@@ -138,6 +138,7 @@ defmodule TdAuthWeb.SessionController do
 
   def create_nonce_session(conn, "saml", json) do
     params = Jason.decode!(json)
+    Map.keys(params)
 
     [saml_response, saml_encoding] =
       ["SAMLResponse", "SAMLEncoding"]
@@ -190,7 +191,8 @@ defmodule TdAuthWeb.SessionController do
 
   defp authenticate_using_saml_and_create_session(conn, saml_response, saml_encoding) do
     with {:ok, profile} <- SamlWorker.validate(saml_response, saml_encoding),
-         {:ok, user} <- Accounts.create_or_update_user(profile) do
+         {:ok, user} <- Accounts.create_or_update_user(profile, true)
+         do
       create_session(conn, user, nil)
     else
       error ->